refactor: organize modules into layered dirs
