+++ src/core/layout-utils.ts
