+++ tests/app-ui.test.ts
@@ -3,7 +3,7 @@ import React from 'react';
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { App } from '../src/app/app';
-import { getDropzoneStyle, undoLastImport } from '../src/ui-utils';
+import { getDropzoneStyle, undoLastImport } from '../src/ui/ui-utils';
 import { tokens } from 'mirotone-react';
 import { GraphProcessor } from '../src/core/GraphProcessor';
 import { CardProcessor } from '../src/board/CardProcessor';
