+++ tests/boardbuilder-extra.test.ts
@@ -1,5 +1,5 @@
 import { BoardBuilder } from '../src/board/BoardBuilder';
-import { templateManager } from '../src/templates';
+import { templateManager } from '../src/board/templates';
 
 interface GlobalWithMiro {
   miro?: { board?: Record<string, unknown> };
