+++ tests/template.test.ts
@@ -1,4 +1,4 @@
-import { templateManager } from '../src/templates';
+import { templateManager } from '../src/board/templates';
 
 interface GlobalWithMiro {
   miro?: { board?: Record<string, unknown> };
